% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkutils-package.R
\docType{package}
\name{wkutils-package}
\alias{wkutils}
\alias{wkutils-package}
\title{wkutils: Utilities for Well-Known Geometry Vectors}
\description{
Provides extra utilities for well-known formats in the
  'wk' package that are outside the scope of that package.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
