% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkutils-package.R
\docType{package}
\name{wkutils-package}
\alias{wkutils}
\alias{wkutils-package}
\title{wkutils: Utilities for Well-Known Geometry Vectors}
\description{
Provides extra utilities for well-known formats in the
  'wk' package that are outside the scope of that package. Utilities
  to parse coordinates from data frames, plot well-known geometry
  vectors, extract meta information from well-known geometry vectors,
  and calculate bounding boxes are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/wkutils}
  \item \url{https://github.com/paleolimbot/wkutils}
  \item Report bugs at \url{https://github.com/paleolimbot/wkutils/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
